backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
boto3==1.34.34
psycopg[binary,pool]==3.1.18
pgvector==0.2.5
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
httpx==0.26.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1
python-dotenv==1.0.0
sse-starlette==1.8.2
